move all exposure calculations to exposure modules
